rename template->interbit-template & set up for publishing
