+++ packages/interbit-template/interbit.config.js
