+++ packages/interbit-template/interbit.prod.config.js
