+++ packages/interbit-template/keyPair.js
