+++ packages/interbit-template/package.json
@@ -1,8 +1,7 @@
 {
-    "name": "template",
+    "name": "interbit-template",
     "description": "Interbit application template implementing PPC (Public, Private, Control) pattern.",
     "version": "0.4.28",
-    "private": true,
     "main": "src/exports.js",
     "license": "MIT",
     "author": "BTL GROUP LTD",
@@ -20,7 +19,6 @@
         "react-router-bootstrap": "^0.24.4",
         "react-router-dom": "^4.2.2",
         "react-scripts": "1.1.0",
-        "react-styleguidist": "^6.2.4",
         "redux": "3.7.2",
         "redux-devtools-extension": "^2.13.2",
         "redux-saga": "^0.16.0",
