+++ packages/interbit-template/public/favicon.ico
