+++ packages/interbit-template/public/index.html
@@ -28,9 +28,10 @@
   <noscript>
     You need to enable JavaScript to run this app.
   </noscript>
-  <div id="interbit" data-boot-react-app="templatePublic" data-peer-hints="ib-dev----master.herokuapp.com" data-chain-id-template-public="32b9365f325da189439f6d453593cf4e0a5a06ead7e6bbecc1d0814c578df452">
+  <div id="interbit" data-boot-react-app="templatePublic" data-peer-hints="localhost:5000" data-chain-id-template-public="2b1020a43670aed4d6d116d2a3226197529dc7f51e6c2ffe68a496539e24c1bc">
   </div>
   <div id="root"></div>
+
