+++ packages/interbit-template/public/manifest.json
