+++ packages/interbit-template/src/App.js
