+++ packages/interbit-template/src/adapters/privateChainAdapter.js
