+++ packages/interbit-template/src/adapters/publicChainAdapter.js
