+++ packages/interbit-template/src/components/LinkedCovenant.js
