+++ packages/interbit-template/src/constants/chainAliases.js
