+++ packages/interbit-template/src/constants/navigation.js
