+++ packages/interbit-template/src/constants/paths.js
