+++ packages/interbit-template/src/containers/CompleteCAuth.js
