+++ packages/interbit-template/src/containers/ExploreChain.js
