+++ packages/interbit-template/src/containers/InteractiveChains.js
