+++ packages/interbit-template/src/index.js
