+++ packages/interbit-template/src/interbit/.eslintrc
