+++ packages/interbit-template/src/interbit/control/index.js
