+++ packages/interbit-template/src/interbit/control/package.json
