+++ packages/interbit-template/src/interbit/private/index.js
