+++ packages/interbit-template/src/interbit/private/package.json
