+++ packages/interbit-template/src/interbit/public/index.js
