+++ packages/interbit-template/src/interbit/public/package.json
