+++ packages/interbit-template/src/redux/exploreChainReducer.js
