+++ packages/interbit-template/src/redux/index.js
