+++ packages/interbit-template/src/registerServiceWorker.js
