+++ packages/interbit-template/src/tests/interbit/controlCovenant.test.js
