+++ packages/interbit-template/src/tests/interbit/privateCovenant.test.js
