+++ packages/interbit-template/src/tests/redux/exploreChainReducer.test.js
