+++ packages/interbit-template/static.json
